+++ packages/f-footer/src/components/tests/LinkList.test.js
@@ -26,8 +26,7 @@ describe('LinkList component', () => {
             window.dispatchEvent(new Event('resize'));
         });
 
-        // TODO Figure out how to test with CSS modules
-        xit('should be in a collapsed state', () => {
+        it('should be in a collapsed state', () => {
             // Arrange & Act
             const linkListWrapper = wrapper.find('[data-js-test="linkList-wrapper"]');